+++ src/features/usage-consolidation-accordion/usage-consolidation-content-package.js
@@ -9,6 +9,7 @@ import SummaryTable from './summary-table';
 import {
   costPerUse as costPerUseShape,
   entityTypes,
+  costPerUseTypes,
 } from '../../constants';
 
 const propTypes = {
@@ -17,6 +18,14 @@ const propTypes = {
 };
 
 const UsageConsolidationContentPackage = props => {
+  const {
+    cost,
+    costPerUse,
+    usage,
+  } = props.costPerUseData.data[costPerUseTypes.PACKAGE_COST_PER_USE]?.attributes?.analysis;
+
+  const noCostPerUseAvailable = !cost && !costPerUse && !usage;
+
   const customProperties = {
     columnMapping: { cost: 'ui-eholdings.usageConsolidation.summary.packageCost' },
   };
@@ -27,6 +36,8 @@ const UsageConsolidationContentPackage = props => {
         id="packageUsageConsolidationSummary"
         entityType={entityTypes.PACKAGE}
         customProperties={customProperties}
+        noCostPerUseAvailable={noCostPerUseAvailable}
+        costPerUseType={costPerUseTypes.PACKAGE_COST_PER_USE}
         {...props}
       />
       <MultiColumnList